requirements.txt
alembic==1.7.5
aniso8601==9.0.1
astroid==2.8.6
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
coverage==6.1.2
coveralls==3.3.1
docopt==0.6.2
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.27
platformdirs==2.4.0
protobuf==3.19.1
pylint==2.11.1
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
